Fix death dim camera

+++ src/main/java/legend/game/combat/environment/CameraOctParamCallback.java
@@ -2,5 +2,5 @@ package legend.game.combat.environment;
 
 @FunctionalInterface
 public interface CameraOctParamCallback {
-  void accept(float x, float y, float z, int ticks, int e, int f, int stepType, int scriptIndex);
+  void accept(float x, float y, float z, int ticks, int e, float f, int stepType, int scriptIndex);
 }
